feat: Add ImageSpec::set_cicp functions and oiiotool --cicp option
The ImageSpec::set_cicp functions are convenience functions for setting the uint8[4] oiio:CICP attribute.

The oiiotool `action_cicp` function works similarly to `action_iscolorspace`, in that it uses a custom OiiotoolOp to modify the top image's spec with the aforementioned set_cicp function.

Upshot: oiiotool now has a `--cicp` option, which makes it easy to add, modify, or remove CICP metadata.
Signed-off-by: Zach Lewis <zachcanbereached@gmail.com>